+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix ig.ENTITY.NPC#postActionUpdate getting overriden when server is off
 - Fix map menu player markers not getting drawn on some maps
 - Fix map menu player markers persisting after restarting the server
+- Fix not all vars getting linked to map/client instance
 
 ## [0.5.7] 2025-11-10